requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
torch>=1.9.0
monai>=1.0.0
tqdm>=4.62.0
transformers>=4.20.0
scikit-image>=0.19.0
open_clip_torch>=2.0.0
torchvision>=0.10.0
Pillow>=8.0.0
scikit-learn>=0.0
wandb>=0.12.0